refactor(sync): one commit ingest, one subscribe auth check
Fold Flutter WS apply_commit_json into ingest_commit_json via
CommitIngestOpts::replica() (signature on, rights/timestamp off). Hub and
peer keep enforcing rights. Named presets replace repeated struct
literals.

Subscribe, SubscribeDrive, and SubscribeQuery share authorize_read so
the three check_read copies cannot drift.

Measure: 2333 → 2257 production non-blank in scope; largest file
engine.rs 1084 → 918.
