+++ lib/src/sync/engine.rs
@@ -17,6 +17,8 @@ use crate::{Db, Storelike};
 
 use super::protocol;
 
+pub use super::ingest::{ingest_commit_json, CommitIngestOpts};
+
 /// Process a single v2 binary frame. Returns response frames to send back.
 /// This is the transport-agnostic entry point — used by WebSocket, Iroh, etc.
 pub async fn handle_frame(
@@ -288,175 +290,6 @@ pub async fn handle_frame(
     }
 }
 
-/// Policy knobs distinguishing a hub ingesting a client's commit from a peer
-/// replica ingesting another peer's commit. The validation *core* (signature,
-/// schema, timestamp, signer rights) is identical in both roles.
-pub struct CommitIngestOpts {
-    /// Transport/source identity for echo suppression by the hub's commit
-    /// monitor. Peers have no commit-monitor fanout, so `None` there.
-    pub source_id: Option<String>,
-    /// Hub semantics: reject commits whose Loro ops are concurrent with stored
-    /// state (client doc wasn't seeded from this node). Off between peers,
-    /// where concurrent writes are expected.
-    pub validate_loro_causality: bool,
-    /// Hub semantics: reject commits for subjects this node cannot own. Off
-    /// for peer replicas — hosting subjects the node does not own is what
-    /// replication is.
-    pub enforce_subject_ownership: bool,
-    /// Peer-transport semantics: hold the importing flag while applying so the
-    /// live push loop doesn't rebroadcast the commit back to live peers (the
-    /// sender included). Off on the hub, where WS fanout is suppressed
-    /// per-source via `source_id` and Iroh live peers SHOULD receive the
-    /// update.
-    pub suppress_live_echo: bool,
-    /// Origin used to resolve `internal:/` subjects in the response JSON-AD.
-    /// `None` falls back to the store's base domain.
-    pub response_origin: Option<String>,
-}
-
-/// Ingest a signed JSON-AD `COMMIT`, returning the server-created commit
-/// resource as JSON-AD. This is the single implementation shared by the
-/// server's HTTP/WS commit application and peer-transport `COMMIT` frames
-/// (see [`CommitIngestOpts`] for what differs between the two roles).
-///
-/// Signature, schema, and signer-rights validation always run — the commit is
-/// a self-authorizing certificate, so those checks (not a connection's AUTH
-/// identity) are the authority. What varies is domain-ownership enforcement,
-/// Loro-causality enforcement, live-echo suppression, and source-id-based
-/// echo suppression, all controlled by `opts`.
-pub async fn ingest_commit_json(
-    store: &Db,
-    commit_json: &str,
-    opts: &CommitIngestOpts,
-) -> crate::errors::AtomicResult<String> {
-    // Reject commits with deprecated set/push/remove fields — use loroUpdate instead.
-    if commit_json.contains("\"https://atomicdata.dev/properties/set\"")
-        || commit_json.contains("\"https://atomicdata.dev/properties/push\"")
-        || commit_json.contains("\"https://atomicdata.dev/properties/remove\"")
-    {
-        return Err(
-            "Commits with `set`, `push`, or `remove` fields are no longer accepted. Use `loroUpdate` instead."
-                .into(),
-        );
-    }
-
-    let incoming_commit_resource =
-        crate::parse::parse_json_ad_commit_resource(commit_json, store).await?;
-    let incoming_commit = crate::commit::Commit::from_resource(incoming_commit_resource)?;
-
-    // Log incoming commit details for debugging
-    if let Some(loro_bytes) = &incoming_commit.loro_update {
-        let doc = crate::loro::AtomicLoroDoc::new();
-        if doc.import_update(loro_bytes).is_ok() {
-            let props = doc.get_all_properties();
-            let prop_summary: Vec<String> = props
-                .keys()
-                .map(|k| k.rsplit('/').next().unwrap_or(k).to_string())
-                .collect();
-            tracing::info!(
-                subject = %incoming_commit.subject,
-                signer = %incoming_commit.signer,
-                properties = ?prop_summary,
-                loro_bytes = loro_bytes.len(),
-                "Incoming commit"
-            );
-        }
-    } else {
-        tracing::info!(
-            subject = %incoming_commit.subject,
-            destroy = ?incoming_commit.destroy,
-            "Incoming commit (no loroUpdate)"
-        );
-    }
-
-    if opts.enforce_subject_ownership {
-        let is_internal = incoming_commit.subject.is_internal();
-        let is_did = incoming_commit.subject.is_did();
-        let matches_base = if let Some(base) = store.get_base_domain() {
-            incoming_commit.subject.as_str().contains(&base)
-        } else {
-            false
-        };
-
-        // Fallback: if it's a local path like http://localhost/ or https://atomicdata.dev/
-        // and it matches the current request's Host, we should also allow it.
-        let is_local_path =
-            !is_did && !is_internal && incoming_commit.subject.as_str().ends_with('/');
-
-        if !is_internal && !is_did && !matches_base && !is_local_path {
-            return Err(
-                "Subject of commit should be sent to other domain - this store can not own this resource."
-                    .into(),
-            );
-        }
-    }
-
-    let signer = incoming_commit.signer.clone();
-    let signer_pure = signer.pure_id();
-
-    // Ensure the agent exists before applying the commit.
-    // This is important because the commit might be editing the agent itself.
-    // Run unconditionally on both roles: a commit rejected later by
-    // `apply_commit` still leaves this auto-created agent resource behind —
-    // accepted hub behavior, now shared with peer ingestion.
-    let is_self_creating_agent =
-        incoming_commit.subject.is_agent_did() && incoming_commit.subject == signer;
-
-    if signer.is_agent_did()
-        && !is_self_creating_agent
-        && store.get_resource(&signer).await.is_err()
-    {
-        let mut new_agent = crate::Resource::new_instance(crate::urls::AGENT, store).await?;
-        new_agent.set_subject(signer_pure.clone());
-        if let Some(pk) = signer.as_str().strip_prefix("did:ad:agent:") {
-            new_agent
-                .set_string(crate::urls::PUBLIC_KEY.into(), pk, store)
-                .await?;
-        }
-        new_agent.save_locally(store).await?;
-        tracing::info!("Auto-created agent resource for {}", signer_pure);
-    }
-
-    let commit_opts = crate::commit::CommitOpts {
-        validate_schema: true,
-        validate_signature: true,
-        // Timestamp validation bounds replay: without it, a captured signed
-        // destroy commit could be replayed unboundedly later (e.g. after the
-        // subject was legitimately recreated). Peers therefore need
-        // roughly-sane clocks — the same requirement AUTH already imposes.
-        validate_timestamp: true,
-        validate_rights: true,
-        // https://github.com/atomicdata-dev/atomic-server/issues/412
-        validate_previous_commit: false,
-        // Reject commits whose Loro ops are concurrent with stored state
-        // (i.e. the client's doc wasn't seeded from the server). Without this,
-        // LWW silently drops the client's write. For P2P sync use a path that
-        // leaves this off — concurrent writes are expected there.
-        validate_loro_causality: opts.validate_loro_causality,
-        validate_for_agent: Some(signer.to_string()),
-        update_index: true,
-        source_id: opts.source_id.clone(),
-    };
-
-    let base_domain = store.get_base_domain();
-
-    let response = if opts.suppress_live_echo {
-        // Applying a remote peer's commit must not rebroadcast to live peers
-        // (the sender included) — mirrors `ws_apply::apply_commit_json`'s
-        // suppression of the same echo via the live push loop.
-        super::ws_apply::set_importing(true);
-        let result = store.apply_commit(incoming_commit, &commit_opts).await;
-        super::ws_apply::set_importing(false);
-        result?
-    } else {
-        store.apply_commit(incoming_commit, &commit_opts).await?
-    };
-
-    let origin = opts.response_origin.as_deref().or(base_domain.as_deref());
-    let json = response.commit_resource.to_json_ad(origin)?;
-    Ok(json)
-}
-
 /// Apply a JSON-AD `COMMIT` received over a peer transport, returning the
 /// server-created commit resource as JSON-AD (the `COMMIT_OK` payload).
 ///
@@ -474,21 +307,9 @@ pub async fn ingest_commit_json(
 /// to another domain): a peer replica legitimately hosts subjects it doesn't
 /// own — that's what replication is — so no such gate applies here.
 async fn apply_peer_commit(store: &Db, commit_json: &str) -> crate::errors::AtomicResult<String> {
-    ingest_commit_json(
-        store,
-        commit_json,
-        &CommitIngestOpts {
-            source_id: None,
-            validate_loro_causality: false,
-            enforce_subject_ownership: false,
-            suppress_live_echo: true,
-            response_origin: None,
-        },
-    )
-    .await
+    ingest_commit_json(store, commit_json, &CommitIngestOpts::peer()).await
 }
 
-/// Collects all resource subjects belonging to a drive via BFS on parent relationships.
 /// Collects all resource subjects belonging to a drive via BFS on parent relationships.
 /// Returns pure_id() strings (no query params/drive hints) to match LoroSnapshot keys.
 pub async fn collect_drive_subjects(
